+++ src/server/ws.rs
@@ -112,7 +112,7 @@ fn get_current_state() -> Option<String> {
     if let Some(current) = &store.current_session {
         total_keystrokes += current.keystrokes;
         total_clicks += current.mouse_clicks;
-        total_duration += current.duration_secs() as i64;
+        total_duration += current.duration_secs();
         unique_apps.insert(current.process_name.clone());
     }
